requirements.txt
streamlit
duckduckgo-search
requests
beautifulsoup4
spacy
reportlab
python-pptx
tqdm
joblib
pydantic
pytest
black
flake8
ddgs
matplotlib

